server/ocpp_cs/requirements.txt
asyncio
websockets
ocpp
pyhumps
swagger-ui-py
aiohttp
aiohttp_sse
aiohttp_cors